Utilize Docker Compose + env variables for deploy
Application is fully docker-ized now, with a PostgreSQL image that will
serve as a database. Environment variables ensure connection. Next step
is to ensure the appropriate tables exist / are created on the DB on
connect.

The use of environment variables could be prettier, but Go would
require me to go to pgdbGO and update the PGConnection struct with a
setter method and I'd rather not at this point in time.